# Speed up `filter_run_end_array`